pyproject.toml
[tool.poetry]
name = "python-todo-app"
version = "0.1.0"
description = ""
authors = ["Kalle Ilves <kalle.ilves@helsinki.fi>"]

[tool.poetry.dependencies]
python = "^3.6"
python-dotenv = "^0.15.0"
invoke = "^1.4.1"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
autopep8 = "^1.5.4"
pytest = "^6.1.2"
pytest-dotenv = "^0.5.2"
coverage = "^5.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
